pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "confluence-markdown-exporter"
version = "3.2.0"
description = "A tool to export Confluence pages to Markdown"
keywords = ["confluence", "atlassian", "markdown", "export", "convertion", "download"]
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Sebastian Penhouet" }
]
requires-python = ">= 3.8"
dependencies = [
    "atlassian-python-api==3.41.21",
    'jmespath',
    'markdownify',
    'pydantic-settings',
    'pyyaml',
    'questionary',
    'tabulate',
    'tqdm',
    'typer',
    'python-dateutil',
    "lxml>=6.0.2",
]


[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.11.13",
]

[project.urls]
Homepage = "https://github.com/Spenhouet/confluence-markdown-exporter"
Documentation = "https://github.com/Spenhouet/confluence-markdown-exporter/blob/main/README.md"
Source = "https://github.com/Spenhouet/confluence-markdown-exporter"
Tracker = "https://github.com/Spenhouet/confluence-markdown-exporter/issues"

[project.scripts]
confluence-markdown-exporter = "confluence_markdown_exporter.main:app"
cf-export = "confluence_markdown_exporter.main:app"

[tool.hatch.build.targets.wheel]
packages = ["confluence_markdown_exporter"]

[tool.ruff]
# Exclude a variety of commonly ignored directories. This means Ruff will not lint or format files with these names
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]

indent-width = 4 # each indent is 4 spaces, equivalent to using "tab" 
line-length = 100 # max no of characters in a line. Black default is 88 characters
target-version = "py310" # Assumes Python 3.10 and above

[tool.ruff.lint]
select = [
  "A", # flake8-builtins
  "B", # flake8-bugbear
  "D", # pydocstyle
  "E", # pycodestyle errors
  "F", # pyflakes
  "G", # flake8-logging-format
  "I", # isort
  "N", # pep8-naming
  "S", # flake8-bandit
  "W", # pycodestyle warnings
  "C4", # flake8-comprehensions
  "EM", # flake8-errmsg
  "PD", # pandas-vet
  "PL", # Pylint
  "UP", # pyupgrade - auto-upgrade syntax for current version of Python
  "ANN", # flake8-annotations
  "BLE", # flake8-blind-except
  "C90", # McCabe complexity checker
  "ERA", # eradicate - removes commented out code
  "FBT", # flake8-boolean-trap
  "FLY", # flynt
  "ICN", # flake8-import-conventions
  "LOG", # flake8-logger
  "NPY", # numpy-specific rules
  "PGH", # pygrep-hooks
  "PIE", # flake8-pie
  "RET", # flake8-return
  "RSE", # flake8-raise
  "SIM", # flake8-simplify
  "RUF", # ruff-specific rules
  "TCH", # flake8-type-checking
  "TID", # flake8-tidy-imports
  "TRY", # tryceratops
  "ASYNC", # flake8-async
  "PT", # flake8-pytest-style
  "FAST", # FastAPI,
  "T20", # flake8-print
  "ARG", # flake8-unused-arguments
  "PTH", # flake8-use-pathlib
  "PERF", # Perflint
  "FURB", # refurb
]

ignore = [
  "W191", # lint rule that may clash with Ruff Formatter: tab-indentation
  "E111", # lint rule that may clash with Ruff Formatter: indentation-with-invalid-multiple
  "E114", # lint rule that may clash with Ruff Formatter: indentation-with-invalid-multiple-comment
  "E117", # lint rule that may clash with Ruff Formatter: over-indented
  "D206", # lint rule that may clash with Ruff Formatter: indent-with-spaces
  "D300", # lint rule that may clash with Ruff Formatter: triple-single-quotes      
  "D1", # ignore this to match google docstring convention
  "G004", # ignore this to allow f-strings in logging
  "UP015", # ignore this to allow "with open" statements to have modes explicitly stated 
  "SIM102", # ignore this to avoid changing nested if statements to single if statements, potentially confusing
  "ERA001", # ignore this to keep commented out lines while functionality is not implemented (configs/logos)
  "PERF203", # ignore this as this often is intentional.
  "ARG002", # Many methods in this project share the same signature, independent of variable usage.
  "PLC0415", # Allow lacy loading of imports
]

fixable = ["ALL"] # Allow fix for all enabled rules (when using "Fix all" or when `--fix` is provided to ruff check in CLI)
unfixable = ["F401"] # disable autofix for unused-imports

dummy-variable-rgx = "^(_+\\w*)$" # Allow unused variables when underscore-prefixed
flake8-bugbear.extend-immutable-calls = [
  "fastapi.Depends",
  "fastapi.Query",
] # Allow default arguments like, e.g., `data: List[str] = fastapi.Query(None)`  
pycodestyle.max-doc-length = 100 # max line-length for docstrings
pydocstyle.convention = "google" # docstring convention. Options: "google", "numpy", or "pep257"
pylint.max-args = 10 # max no of args in a function

[tool.ruff.lint.isort]
known-first-party = ["airamed", "main"]
force-single-line = true # force each import to be in its own line

[tool.ruff.format]
docstring-code-format = true # Enable auto-formatting of code examples in docstrings. Markdown, reStructuredText code/literal blocks and doctests are all supported
docstring-code-line-length = "dynamic" # Set line length limit used when formatting code snippets in docstrings. This only has an effect when the `docstring-code-format` setting is enabled
indent-style = "space" # indent with spaces, rather than "tab"
line-ending = "lf" # options: "auto", "lf", "cr-lf", "native"
quote-style = "double" # Use double quotes as voted by majority
skip-magic-trailing-comma = false # respects magic trailing commas

# Ignore S101 (assert) in all test files
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
  "S101",    # Assert in tests is expected
  "S110",    # try-except-pass detected
  "FBT001",  # Often conflicts in tests
  "PLR2004", # Magic numbers are acceptable in tests
]
